refactor: move observe setup to util and add recovery in coordinator
- Add async_setup_observe() to util.py with all CoAP observation
  callbacks (scenes, targets, sensors, radar, energy, config)
- async_setup_observe sets observe_active=True on success,
  False on PyTewkeObserveError (so the coordinator can retry)
- Also triggers initial scene processing for data already fetched
  during the first coordinator refresh
- Add observe_active: bool = False field to TewkeData
- __init__.py now calls async_setup_observe() instead of defining
  callbacks inline; removes the PLR0915 noqa suppressor
- coordinator._async_update_data checks observe_active and calls
  async_setup_observe to re-establish observations if they are down
