+++ custom_components/tewke/data.py
@@ -24,3 +24,4 @@ class TewkeData:
     coordinator: TewkeCoordinator
     scene_control_types: dict[str, str]
     pending_scenes: dict[str, Scene] = field(default_factory=dict)
+    observe_active: bool = False
